+++ plugins/sidebar-panel/index.js
@@ -4,7 +4,7 @@ import {
 } from "../../common/plugin-element-cache";
 import { getCtdSettings } from "../../common/settings-parser";
 import pluginInfo from "../../plugin-manifest.json";
-import { clearConnections } from "../field-config/websockets";
+import { clearConnections } from "../../common/websockets";
 import { createPanelElement, updatePanelElement } from "./panel-elements";
 
 export const handlePanelPlugin = (
@@ -29,7 +29,13 @@ export const handlePanelPlugin = (
   const spaceId = getSpaceId();
 
   const objectData = { ...contentObject, ...formik.values };
-  updatePanelElement(pluginContainer, settingsForCtd, objectData, spaceId, create);
+  updatePanelElement(
+    pluginContainer,
+    settingsForCtd,
+    objectData,
+    spaceId,
+    create,
+  );
 
   return pluginContainer;
 };