+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
+## [0.3.0]
+### Fixed
+* storing nested objects in yjs doc #26528
+
+### Added
+* translation field on translations add #26528
+
 ## [0.2.0]
 ### Changed
 * URL to live preview #26528